requirements.txt
certifi==2019.3.9
chardet==3.0.4
future==0.17.1
idna==2.7
python-telegram-bot==12.0.0b1
PyYAML==5.1
requests==2.19.1
urllib3==1.25.2
